pyproject.toml
[project]
name = "atcoder-gacha-bot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "shunsei", email = "" }
]
dependencies = [
    "python-dotenv>=1.0.0",
    "ipykernel>=6.27.1",
    "requests>=2.31.0",
    "slack_bolt>=1.18.1",
]
readme = "README.md"
requires-python = ">= 3.10"

[project.scripts]
hello = "atcoder_review_bot:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/atcoder_review_bot"]
